fix: style change from inline to scss
