+++ src/storybook/components/related-components/descriptions/virtualized-grid-description/virtualizedGrid.module.scss
@@ -1,0 +1,13 @@
+.virtualizedGridItem{
+    display: flex;
+    align-items: center;
+    overflow-x: hidden;
+    width: 230px;
+    height: 100%;
+}
+
+.virtualizedGridItemChild
+{
+    width: inherit;
+    height: inherit;
+}
